pyproject.toml
[project]
name = "aquiles-backend"
version = "0.1.0"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "sqlalchemy>=2",
  "psycopg2-binary",
  "python-jose[cryptography]",
  "passlib[bcrypt]",
  "httpx",
  "pydantic-settings",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
